requirements.txt
polygon-api-client
pandas
numpy
plotly
scipy